[UPDATE] Added documentation for comicsByMonth, moved months array to data folder
