+++ helpers/comicsInMo.js
@@ -1,7 +1,6 @@
-import months from './months';
+import months from '../data/months';
 
 /**
- * 
  * @param {array} comicsByYr  an array of comics for the year
  * @param {string} mo a month string eg "jun"
  * @returns {array} an array of comics for the month
